+++ terra/contracts/cw20-wrapped/src/state.rs
@@ -9,7 +9,7 @@ pub const KEY_WRAPPED_ASSET: &[u8] = b"wrappedAsset";
 // Created at initialization and reference original asset and bridge address
 #[derive(Serialize, Deserialize, Clone, Debug, PartialEq, JsonSchema)]
 pub struct WrappedAssetInfo {
-    pub asset_chain: u8,        // Asset chain id
+    pub asset_chain: u16,        // Asset chain id
     pub asset_address: Binary, // Asset smart contract address on the original chain
     pub bridge: CanonicalAddr,  // Bridge address, authorized to mint and burn wrapped tokens
 }
